+++ src/libraries/System.Private.CoreLib/src/System/Collections/Generic/ValueListBuilder.cs
@@ -107,7 +107,7 @@ namespace System.Collections.Generic
 
             int pos = _pos;
             Span<T> span = _span;
-            if ((ulong)(uint)pos + (ulong)(uint)length <= (ulong)(uint)span.Length) // same guard condition as in Span<T>.Slice on 64-bit
+            if ((ulong)(uint)pos + (ulong)(uint)length <= (ulong)span.Length) // same guard condition as in Span<T>.Slice on 64-bit
             {
                 _pos = pos + length;
                 return span.Slice(pos, length);
